Fix param copying routines to respect different gs_memory_t.
We have a routine gs_param_read_items that takes a parameter list,
a structure, and a gs_param_item_t list. Each gs_param_item_t gives
a parameter to read, and an offset to store it in the structure.

This is used by various routines to bulk read parameter values, and
to store them in (normally) device structures.

This works fine, when we are reading simple parameter values like
ints or floats. When reading more complex values, like Strings or
arrays, the existing code just copies the pointers to the underlying
data, rather than copying the data itself.

This has 2 problems. Firstly, it relies on the copies of the data in
the param list persisting for longer than the copied structure that
points into this data. Secondly, if data is copied from a param list
into a structure, then both the param list and the structure must
share the same memory pointer to ensure that the garbage collector
does not get confused.

In most places within the code we aren't copying 'complex' param
values (i.e. we aren't copying arrays, so this isn't an issue). The
exceptions to this are the JPEG Encoding stream, and pdfwrite/ps2write.

We therefore extend gs_param_read_items to take an addition parameter,
a memory pointer, and tweak its operation. If we ever try to copy a
dictionary or 'general' array type, we exit with a rangecheck.
Nowhere in the code appears to do this, so no changes there.

If the memory pointer is NULL, we perform exactly as we do now. We
update all the calling positions other than the pdfwrite/ps2write
cases to pass NULL, so no changes there.

In the pdfwrite/ps2write cases, we pass the memory pointer for
the device structure into which the data is being copied. For
param types that need it, the new code allocates memory with the
correct memory pointer and copies the values into it.