native font map from font file scan
First pass at FONTDIR scanning...

Handles Type 1 and TTF/TTC/OTTO files.

CFF still to do. And probably some tidying up.

Slightly rejig font file scanning.

Previously, the code used the magic number to decide which file scanning
function to call, which would attempt to find the font name in the file, and
return it, so the caller could add the name/path to the native font map.

But that approach was messy when the ability to deal with TrueType collections
was implemented.

So, this tweaks it so the scanning function also does the adding of the font
mapping to the dictionary.