GhostPDF - prevent accidental triggering of 'REPAIRED_KEYWORD'
This came up with Robin's work on N-up; using ps2write a shading was
converted to an image, and then discovered it was entirely clipped.

This led to the return code from the bounding box retrieval being
passed all the way back up to the PDF interpreter. The retrieval return
code is slightly unusual in that instead of 0 being success, it means
there was no box, and 1 means success.

Because the '1' made it all the way back to the interpreter we
mis-interpreted it as a 'REPAIRED_KEYWORD' which caused a crash.

This was fixed in the pdfwrite code, but it's impossible to be certain
there are no other possible cases, so this commit makes sure we never
return a value > 0 from pdfi_interpret_stream_operator() unless we
really have repaired a keyword.