pdfwrite - Error out on degenerate pattern matrix
Bug #704395 "Ghostscript - Bad pattern produced with Nup
 and pdfwrite"

There's more explanation in the bug report, and I don't want to
recapitulate it here. Basically if we instantiate a Pattern and the
saved CTM (which is a combination of the matrix argument to makepattern
and the CTM at the time makepattern was executed) is degnerate then we
now return an error.

Previously we would write out a degenerate matrix to the final PDF or
PostScript file which would then fail. It seems better to error before
producing the output.