Fix Bug 704330: NupControl and clist problems to pkmraw.
Files that have a page with transparency followed by a page that does
not use transparency in the same nest that are using the clist, such as
the file for this bug and tests_private/comparefiles/Bug696687.pdf are
examples.

The problem is that the clist writer uses the 'clist_color_info' to
determine the depth, etc. when writing colors into the clist, and a page
that uses transparency changes it, but did not restore it when the pdf14
device was popped (disabled). The fix is simple -- restore this as well
as the color_info for the clistdev.