bbox device - don't short-circuit fills in stand-alone mode
Commit 378f362b25dc51d92e9668420a0195e8253123c2 updated the bbox device
to deal with the changed prototype for the nullpage device, but an
optimisation was misplaced.

This led to returning from the bbox device without scan-covnerting the
path and determining the bounds, if the bbox device was in stand-alone
mode.

This commit moves the test for stand-alone mode inside the text for the
bbox of the path being entirely within the current bbox, in which case
we don't need to process it for the bbox device, and if there is no
target device (fill_path is NULL) then we can just ignore the path
altogether).

We do also need to avoid drawing to the target device if it is NULL
(stand-alone) so move that into the check for tdev == NULL and refactor
the code a little.