GhostPDF - performance enhancement with memory streams
When we create an in-memory stream from a filtered file stream, we do
not know the length of the decompressed stream (PDF does not have a
/Length2 for anything except fonts). So we end up reading the stream
twice, once to determine the length, then allocate memory to hold the
data, then read the stream again, into the buffer.

We were doing the initial read a byte at a time, which turns out to be
slow, this commit reads up to 512 bytes at a time which is noticeably
faster on some files.

Interestingly attempting to avoid the double read by reading into a
buffer, then reallocating and copying as required, turned out to be
significantly slower even than the original code.