GhostPDF - performance enhancement on colour spaces
This commit checks the action of setting the current colour space to a
named colour space, to ensure that it is not already the current colour
space.

This is unfortunately complicated. Because of gsave and grestore, we
must store some object along with the colour space which indicates
the PDF object that it is associated with, so that we can check the
current fill or stroke colour space against the requested space.

Fortunately we already have provision for storing the interpreter
context in the colour space, in order to deal with functions describing
the tint transform to the alternate space.

We take advantage of this to store the PDF interpreter 'name' object
instead, as PDF objects store a pointer to the context anyway, so we
can retrieve the context from there in order to free functions.

However... Not all colour spaces are associated with a PDF name. For
example the Alternate space of a Separation or DeviceN space, and
device spaces. In this case we do actually need to store a pointer to
the PDF interpreter context. But we need to be able to tell which one
we have stored.

The simplest way to deal with this is to turn the PDF interpreter
context into a PDF interpreter object, of a sort. We won't really
reference count it, or consult it's flags, but we can check the object
type to see what kind of object it is, and use the context pointer
without concerns.

The remaining ugliness is storing a pointer to the current 'name' in the
PDF interpreter context when we are asked to set a named space, just
so that at the bottom level we can store it in the colour space (and
count it up!). I'd prefer to do this by passing a parameter but that
requires some replumbing that I don't really want to tackle just yet.

The colour space testing is conservative, if there is any doubt that the
current space is not the same as the requested space, we will build
and set the requested space.

The test file I've been using, page 5 from the file sent by Bruno Voisin
on gs-devel, now runs 2.4 times faster, and is broadly the same speed
as the old PDF interpreter at screen resolution. The remaining pages
seem to be broadly in line as well.