Bug 704410 Overprint/Transparency Fix
If halftone screens are encountered when processing
a file, this can have an effect of setting the
opmsupported (overprint mode supported) variable in
the device color info.  The pdf14 device changes its
color information based upon the group color space
and/or if it is dealing with a softmask.  This fix
ensures that the opmsupported value is set to UNKNOWN
when a colorspace change occurs, forcing the overprint
code to determine the drawn components correctly based
upon the current state.

In addition, a CMYK fill with [0 0 0 0] where overprint
is true and overprint mode is true, is a NOP since
no drawing should be made.  This commit short circuits
the pdf14 mark fill rect operations to simply return if
this is the case.

Finally, make sure the pdfi code disables OPM prior
to doing the shading.