Revert "Fix Bug 703898 commit cc3eebdffa
Adding spot colorants when seen in Type 5 halftones could end up adding
some that are not counted in the PageSpotColors count of PDF files, that
were not counted because they did not occur in any colorspace.

If a Type 5 Halftone is set for a spot color before the colorspace that
uses that spot color, it will be ignored, which still needs to be fixed.
That will be a subsequent commit.