Make ICC profile cache use a consistent memory allocator
The cache contains it's own memory allocator, and we should use that
to allocate and free the cache entries. Prior to this commit we used
the graphics state memory allocator to allocate profile entries in
gsicc_add_cs() and we used the graphics state memory pointer to
free them in rc_gsicc_profile_cache_free() but we used the cache
memory allocator to free entries in gsicc_remove_cs_entry().

We need to be consistent in case the allocator in the graphics state
and the cache are not the same (as may be the case with pdfi running
inside Ghostscript). Since we have an allocator in the cache structure
I'm choosing to use that, also we don't have the graphics state memory
allocator available in gsicc_remove_cs_entry().