GhostPDF - Fix ICC profile caching with colour spaces
Commit c21f5f70f6ef15f3d1ec1975d5510e2c0dd959fd improved the
performance in part by spotting attempts to change colour space
to the same space as the current, and avoiding that.

This led to seg faults with pdfi running under Ghostscript...

To detect duplicate spaces, we added the PDF name object of the
space to the 'interpreter data' stored in the colour space. We
then checked the current name against the name of the desired space.

This caused problems with ICC spaces, because these maintain a
reference to the colour space. If the profile was stored in the
PostScript profile cache (which it would be, because the profile
cache is stored in the grapchis state, which we inherit from the
PostScript interpreter) then it would outlive the PDF interpreter,
and when we came to free the cache it would try to free the colour
space, which would try to free the PDF name object, which had already
been freed.

To work around this we replaced the profile cache in the graphics
state with the profile cache from the PDF interpreter. Now when we
free the PDF profile cache, it frees the colour spaces, and the PDF
names, while they are still valid.

However, we then discovered that the garbage collector could reclaim
the colour spaces. This happens because no GC-visible object declares
it has a reference to the colour space. It doesn't matter that the
colour space has been counted up, the GC frees it. Now the profile
cache in pdfi is pointing to a colour space which has gone away....

To get round this, we create a brand-new profile cache, using GC'ed
memory (well actually the interpreter allocator, which is GC'ed for
Ghostscript). When we switch to pdfi we replace the graphics state
pdfi is using with the one from the PostScript environment, and at
the same time we swap out the profile cache being used by that
graphics state, so it uses our new one. On return to PostScript we
swap them back. This allows the lifetime of the profiles in the
profile cache to be correct, because they are visible to the garbage
collector (the PDF context is a GC-visible object and enumerates its
member profile cache)

When we finalize the PDF context we count down the profile cache which
dereferences all its entries. This effectively frees the colour space
(the memory may not be reclaimed yet) which calls the interpreter
specific cleanup, which frees the (still valid at this point) PDF name
object and sets the pointer to NULL.

The net result is that we have the profiles visible to the garbage
collector during their lifetime, but still have control over the cache,
so we can empty it of the profiles created for pdfi when we close the
PDF interpreter.