Fix Bug 703898: Spot color halftones ignored if not in prior colorspace
If a spot color is in a HalftoneType 5 halftone in a PDF, but a colorspace
has not been set previously that uses the spot color, the halftone will
be ignored. This is because we need the colorant number in order to create
the pgs->halftone->params.multiple that includes that colorant.

The approach it to have the PDF interpreter set the SeparationColorNames
from the pagespotcolors dictionary created at the start of each page.

For the transparency compositor (pdf14) we also need to copy any components
that have been set into the devn_params for the pdf14 device (doing new
allocations).

For the psdcmyk device, the usage of SeparationColorNames results in a
all of those colorants being output in the PSD (previously, only colors
that were actually used were present). This seems to match what the Adobe
"Output Preview" shows (example Bug689492.pdf now shows PANTONE 355 CVC).

NOTE: This will require the NEWPDF interpreter to also emit the device
parameter, SeparationColorNames as is done with this patch to pdf_main.ps