GhostPDF - fix soft-masks with pdfwrite
The pdfwrite device, when encountering a soft mask, stores the ID of
the soft mask in the graphics state. It then uses this to determine
whether a soft mask is in operation, and whether it is the same soft
mask as the current one in the pdfwrite context. When the ID in the
graphics state does not match the current ID, we need to grestore.

The problem here is that we instantiate an ExtGState, which has a SMask
and then perform a pair of fill operations. Both of these are performed
while the soft mask is active.

The pdfi interpreter does a gsave/grestore round the transparency
setup and teardown, which causes the soft_mask_id in the graphics state
to be reset. This causes pdfwrite to emit a grestore before the second
fill with the result that only the first fill takes place inside the
soft mask.

This is actually a pretty poor way to operate, but I'm wary of trying
to fix it in pdfwrite currently. Instead, alter the pdfi code so that we
do the gsave/grestore round the fill/stroke/fillstroke operation but not
around the transparency setup and teardown.

I did try removing the gsave/grestore altogether because it seemed to
me we shouldn't need it, but that caused other files to fail.

This fixes tests_private/comparefiles/Bug689931.pdf and
tests_private/comparefiles/Bug689931a.pdf with pdfwrite