pdfwrite - Improve type 3 font capture with GhostPDF
GhostPDF doesn't need to exit to the interpreter in order to run a
type 3 font's CharProc, so it uses the 'early accumulator' case in
pdf_text_process.

However, that path was previously only used by the PCL interpreter
and in order to deal properly with the HPGL 'stick font' we create
a custom method to temporarily replace the device's get_initial_matrix()
method.

We do not want to do that for PDF type 3 fonts!