pdfwrite - fix patterns in coloured type 3 fonts
Type 3 coloured fonts may use pattern colour spaces for stroke and/or
fill. The problem here is that when accumulating a CharProc for a type 3
font we replace the CTM with a 100x100 scaling, see the comments
in install_PS_charproc_accumulator.

Several places (eg pdf_text_set_cache) undo this scaling, but we were
not reversing it when executing a pattern, which resulted in the pattern
being scaled 100 times too large.

This commit prevents that and fixes two Quality Logic test files with
both GhosScript and GhostPDF. Although this had always been wrong it
hadn't previously been noticed (they are PDF 2.0 test files).

Fixes
/tests_private/pdf/PDF_2.0_FTS/fts_23_2310.pdf
/tests_private/pdf/PDF_2.0_FTS/fts_23_2311.pdf