Bug 704463: Fix pngalpha/clist/pdf14 interaction.
When the clist makes the buffer device to render into, it was
(correctly) calling pngalpha to do this. pngalpha was failing
to give it a dev_spec_op that would claim that it supported
alpha. Because of this, the pdf14 buffered data was being
written onto the device opaquely.

A simple 1 line fix gives the buffer device the same dev_spec_op
handler as the main device. This is just the same as the default,
but with 'supports_alpha' returning true.