Coverity IDs 372275, 373352 and 373353
Arising from the work to detect switching to the current colour space.

In pdf_color.c, don't try and set the colour callback in a colour space
if we failed to create the colour space.

In zpdfops.c, move the freeing of the profile cache into the if clause
that checks if pdfctx was allocated. This solves both 373352 and 373353.
Also change the if(ctx) and if(pdfctx) to explicit checks against NULL.