GhostPDF - scale type 3 fonts when capturing CharProc
This is the duplicate of an old fix with Ghostscript, but the GhostPDF
type 3 font interpretation takes a different path (does not use the
PostScript charproc accumulator) which evades the fix.

Essentially the problem is that when we come to draw the glyph but are
forced to fall back and render to an image, the scaling in play results
in FreeType trying to render to a tiny raster, in the test case a 2x2
pixel bitmap.

This was fixed for PostScript type 3 CharProcs by scaling up the CTM
by a factor of 100 so that there is more scope for rendering. This of
course must be accounted for when drawing the image in the CharProc.

Getting this to work is unfortunately complicated by the various
interpreters and differing font types which means there is not easy
simple path through the pdfwrite code for any given font type. Instead
I've had to add another kludgy flag. One day we should rewrite the
pdfwrite font and text handling.