GhostPDF - check content streams for recursion
Bug #704487 "Crash in pdf14_pop_transparency_mask"

This fuzzed file is badly garbled, but it exhibits two cases of
recursion. This is due to the fact that we must inherit Resources
for a content stream from the Page (or Parent) stream if the stream
itself does not define them.

This makes it impossible to detect recursion before a content stream
is executed.

In the case of this file the page fills a rectangle with a Pattern, Sh1.
That Pattern is defined for the page as object 14. While running the
Pattern stream, it executes a 'gs' with the ExtGState /Gs1, which is
defined in the stream dictionary as object 13.

The ExtGstate has a transparency Group, object 12, which is a Form
XObject. The stream for the form fills a rectangle with a Pattern, Sh1
which is not defined in the stream dictionary (which defines a Pattern
named Sh0). We consult the Parent to find the resource, and eventually
discover a Pattern Resource named Sh1 in the Page Resources.

So we execute the stream for that Pattern, which is object 14, and round
and round we go.

We can easily detect hits by checking the object number of the stream
to be executed against the current stream being executed and the parent
of that stream and so on. If we are trying to execute a stream which we
are already in the midst of executing, then we have recursion. We
already store the parent object in order to find missing Resources so
all we need to do is check.

The second case of recursion is in object 14 itself. After setting the
ExtGState it then fills a rectangle with a Pattern. Again the pattern
is named Sh1 and the stream dictionary has no definition of a Pattern
by that name (it has a /Sh0). This same fix catches that one too.