PDF interpreter - check Pattern PaintProc for recursion
Bug #704487 "Crash in pdf14_pop_transparency_mask"

This fuzzed file is badly garbled, but it exhibits two cases of
recursion. This is due to the fact that we must inherit Resources
for a content stream from the Page (or Parent) stream if the stream
itself does not define them.

This makes it impossible to detect recursion before a content stream
is executed.

In the case of this file the page fills a rectangle with a Pattern, Sh1.
That Pattern is defined for the page as object 14. While running the
Pattern stream, it executes a 'gs' with the ExtGState /Gs1, which is
defined in the stream dictionary as object 13.

The ExtGstate has a transparency Group, object 12, which is a Form
XObject. The stream for the form fills a rectangle with a Pattern, Sh1
which is not defined in the stream dictionary (which defines a Pattern
named Sh0). We consult the Parent to find the resource, and eventually
discover a Pattern Resource named Sh1 in the Page Resources.

So we execute the stream for that Pattern, which is object 14, and round
and round we go.

In the PostScript-based PDF interpreter detecting and correcting this
is quite painful. In fact, the best I can do is to simply note the
object number of the current pattern in a dictionary, and check that
dictionary before beginning any new pattern.

Testing suggests it is possible to reach here without the object number
being set in the pattern dictionary, which means that it is potentially
possible to evade this check. In addition, unlike the pdfi code, this
only addresses Patterns, not Forms or other content streams.

Hopefully this will tide us over until we switch interpreters.