Bug 704439: Deal better with extreme glyph scaling.
In the event that a font is scaled up to such an extent that it would overflow
the 16.16 fixed point representation in FAPI, produce a glyph scaled to the
resolution only (saves underflow!) and scale the points in the outline as we
create the Ghostscript path later on.

Also, take a harder line if we overflow Ghostscript's own fixed point
coordinates.