Bug 704567: Fix problem (SEGVs) with commit b12780fc66
The clist buffer devices need to forward the ret_devn_params from the
target device, and the previous commit did not do this for all cases,
seen with the psdcmykog device.

Move the set of the memory device ret_devn_proc to the forwarding
version to the shared procedure: gx_default_create_buf_device.

This was not seen with the default regression -- it was encountered
with the 'extended' test mode of nightly runs.

TBD: The 'display_create_buf_device' needs to be checked, and probably
needs a similar change if it is in a "DeviceN" configuration.