Bug 703785: Tweaking ft over/underflow workarounds
There was a workaround for underflow conditions when running FAPI with
pdfwrite/ps2write, which caused the glyph code to throw errors, or produce poor
glyph positioning in the vector output - the workaround raised the smallest
Freetype scale we'd allow by an order magnitude.

However, that could cause poor rendering at smallish scaling, with certain
fonts.

The original pdfwrite/ps2write output issues are no longer relevant, and the
errors from the glyph code are resolved by a tweak to the degenerate matrix
matrix handling in the FAPI interface.