mswinpr2 device - use non-ASCII characters in printer name
bug #702290 "mswinpr2 doesn't work if the printer has a name with a German umlaut"

The problem is to do with the way we pass the printer name, and use it
in Windows API calls.

We pass the name as UTF-8, and when opening the printer we convert
that into Unicode and wide character. However when getting the device
capabilities we were leaving the printer name as the UTF8 string, which
doesn't work, the Windows API doesn't understand UTF-8.

Instead use the name returned in the DEVMODE structure, that seems to
work better.