tiffsep - add sep number to filename if sep name matches a process ink
Bug #698881 "Problem with spot color with same name of master color (CMJK) "

In this case the separation is named "YELLOW" which is not the same as
"Yellow" as far as PostScript is concerned because ink names are case-
sensitive.

However, when added to the output filename, not all OS's (in particular
Windows) differentiate filenames based on case, leading to the file
being overwritten.

This commit checks the separation name to see if it matches any of the
process inks (ignoring case) and if it does, it appends the separation
number to the ink name for the purpose of creating the filename.