Bug 703694: Fix gs command line handling of double-quotes.
Double quotes are accepted within @file handling to allow for
quoting whitespace. The old code used to accept them at the
top level too, and was happy to accept them not as the start/end
of params. So:

 test" "file.eps

would have worked. Let's not do that.

This means we can now do things like:

 "somewhat"convoluted.eps"

in an @file to mean the file named:

 somewhat"convoluted.eps

The old code used to also accept quotes for command line parameters.
We no longer allow for them there.

This means both command line parameters and @files can properly
signal any valid filename.