display device - fix Separations
Bug #704576 "Regression: Missing content and spot colors with Windows display -dDisplayFormat=16#a0800"

The problem is that since we retained the erasepage optimisation device
the display device is no longer at the head of the chain of devices.

We do want to walk back up the chain and use the topmost device's
callback, but we need to not corrupt the 'dev' pointer while doing so
as we need that to retrieve the DeviceN information.

This commit just adds a new localvariable and sets it to the head of
the chain. We then use that device pointer to execute the callback. This
also saves us walking back up the chain on every ink (including the
process inks). While it is exceedingly quick, it's still effort we don't
need to do.