Ensure psdcmyk device only uses rendered colorants
The page_spot_colors value is truncated in
devn_put_params if the max_components of the device
is ever going to be exceeded.  We should base our
colorant count for the psdcmyk value off of this,
not num_separations, for PDF.

Also clean up some warnings in the alphabetical
sorting code.