Work around a gcc (10.3.0) optimiser bug
The optimiser appears to elide a variable equals zero conditional, meaning we
then try to divide by that variable, causing a division-by-zero sigfpe.

Instead of comparing to the value "0", define a volatile variable whose value
is zero, and compare to that - making it "volatile" prevents the optimiser
dropping the test.