PDF interpreter - reset /File member of pattern after executing PaintProc
Bug #704388 "Rasterizing removes color in some shapes"

The problem only occurs if we set a Pattern space, and a Pattern, and
the PaintProc for the pattern exceeds 65535 bytes, and the pattern
usage is :

Set the /Pattern space,
set a Pattern instance,
fill a path,
gsave
set a new Pattern instance,
fill a path
grestore
fill a path.

This causes us to create a /File entry for the first Pattern instance
which is stored in the pattern dictionary. We then execute the PaintProc
which uses the stored /File to draw the contents.

We then switch to a new Pattern, execute that similarly and then switch
back to the original pattern.

At this point we need to run the original pattern PaintProc again, so
we retrieve the /File entry from the pattern dictionary and again use it
to draw the pattern content.

Now, depending on the size of the stream used to draw the pattern
contents, we either store it as a string (<64KB) or we turn it into a
ReusableStream (>=64KB, or 0 which is illegal but happens and means we
can't tell if the stream will fit in a 64KB string).

The code to execute the pattern contents puts a SubFileDecode onto the
/File entry to create a file object to execute. If the /File object is
a string this effectively creates a file object anew each time. If it's
a file object already then the two get chained.

Strings sources always work, but it turns out that ReusableStream files
don't. This is why Robin was able to switch between working and non-
working cases by changing the stream Length.

This is because the first time we run the PaintProc we read from the
file object until it reaches EOF. For string sources this isn't a
problem as we'll make a new file next time round. For a file source,
however, this means the underlying file source *also* reached EOF.

When we try to run it a second time, we immediately hit EOF and exit
without drawing anything (this is also not an error, empty files or
PaintProc are valid).

This commit fixes the problem by looking at the /File object after we
run the PaintProc, and if its a file (as opposed to string) type, then
we reset the fileposition to 0, so we can run it again.

pdfi, of course, does not suffer from this problem.

We could also fix this by not applying the 64KB limit to the /File and
always create a string source.