pdfi - report errors when integrated with Ghostscript
bug #704624 "NEWPDF doesn't throw an error on reading content stream [REGRESSION]"

This isn't actually a regression, because previously you couldn't have
used 'NEWPDF'.

The new PDF interpreter does not report errors as they occur (except in
debug configurations); instead it saves up all the errors and warnings
and reports them at the end. This means that we only need to report each
category of error once, because people were complaining that broken
files could lead to a deluge of error messages.

However, I forgot to add the code to the PostScript integration to
report the errors at the end of the file.

This commit adds that in.