GhostPDF - don't flag parentheses errors when they aren't errors
I keep forgetting about this one. Balanced parentheses don't have to be
escaped, which actually makes it rather hard to detect a problem if
a string does contain an unbalanced '('.

Anyway, this commit prevents the error being flagged inappropriately
and only sets it if we run out of bytes to read and have an unbalanced
number of parentheses.