GhostPDF - generate SeparationColorNames and send to device
Bug #704546 " Bug 703898 requires changes to the NEWPDF (pdfi) interpreter for SeparationColorNames"

The old PostScript-based PDF interpreter was modified to create an array
of names containing each Separation colour found when scanning the PDF
file Resources and send that array to the device.

This commit adds the same functionality to the new C-based PDF
interpreter.

Unfortunately there are no instructions in bug #704546, or the bug for
the PostScript-based interpreter, bug #703898, on how to actually test
this.

This causes a few differences with the psdcmyk device, mostly due to
creating extra spot colours I believe. I'm going to assume for now that
this works (I tried a file with 6 spot colours and it rendered
correctly to tiffsep) and if there is a problem, it'll show up later.