GhostPDF - reset spot capable devices to undo PL setting
The PL layer specifically disables spot colour handling, possibly for
the benefit of the XPS interpreter, the comment in
pl_main_process_options() isn't totally clear.

The only way for us to put that back, and get more than 4 spot plates
out of a spot capable device, is to close the device, reset the value
to -1, and reopen it.

This leads to us closing and opening the device a lot more than we need
to, which is very wasteful, for the future; try and optimise this and
pdf_check so that we don't need to reopen the device so much.