GhostPDF - fix -dUse...Box switches with Ghostscript integration
Bug #704647 "problems with new PDF-Interpreter"

These got broken at some point before the rush to release, and I didn't
notice until after the release. Fix that here, and additionally cater
for the requested Box being empty.

Also check the type of the box and make sure it's an array.

Unlike the old PDF interpreter, if either of these fails we fall back to
the MediaBox and only use the current page size if the MediaBox is
absent, invalid type or has zero width or height.