Add MEMENTO_TRACKREFS option.
When in MEMENTO_LEAKONLY mode, the largest user of time in memento
is when it tracks references. Add an option to allow that to be
avoided.

This time can probably be drastically reduced by using a splay tree
so we can binary search for a block, rather than having to linear
search. A future enhancement.