Bug 704639  Memory leak in halftone creation from clist
When the halftone data is read from the clist, the transfer
function is allocated but was not returned when the function
type was identity.  This problem showed up more so with this
large file which had lots of colorants.   If the transfer function
is none or identity we set the value to NULL.

Thanks to Robin Watts for helping with Memento and tracking
down the problem.