pdfwrite - correct the 2-byte Unicode test
Bug #704478 (redux) "pdfwrite emits incorrect ToUnicode entries"

Commit 	b4e8434defb8e05ea05bb130b92217290efd2fba fixed the generation
of the required tables, but....

The test against TwoByteToUnicode in pdf_simple_font_needs_ToUnicode()
was inverted. If the font has a ToUnicode CMap and uses glyphs from that
CMap which require more than 2 bytes to represent, then we should not
emit a ToUnicode CMap, because we can't correctly do so.

However the code was returning 'true' which meant that it would try to
write out a ToUnicode CMap; exactly the opposite of what was intended.

Fixing that resolves the problem when both quoteright and ffi or similar
ligatures are used in the same file.

We still cannot preserve ToUnicode CMap information which requires more
than 2 bytes to represent. But this is an enhancement, see bug
#704674.