Bug 699209: Continued. Attempt to fix issue seen with C303.BIN
In the PS and PDF world, images can either be imagemasks (in which
case they are filled in the current gstate color), or they can
supply different colors per pixel.

Accordingly, if an image uses the current gstate color, the clist
must write the halftone phase if the color might need it. If an image
does not use the current gstate color, then the halftone phase should
be written (because we can't tell whether all the colors are pure or
not in advance). The current code copes with this just fine.

Unfortunately, in the PCL world, we can be in the unholy situation
where an image can use the current gstate color AND it can use
colors information from the image. Thus we need to send the phase
both in the case where the current gstate color might need it and
when the image is not a masked one.