Bug 704678: Make sure to store obj/generation number in font objects
For Truetype fonts, the lines that copied the object and generation numbers from
the font dictionary object to the new font object had been dropped, thus meaning
we'd no longer cache the font objects, and reuse them.

Reinstate that. And make sure both values are stored for all fonts. And, again
for all fonts, also store the indirect num/generation values - generally not
much use (for these objects), but in the interest of consistency....